Reference

This package doesn't export any symbols.

It is recommended to import as:

import SquareIsingModel as Ising

SquareIsingModel.βc — Constant

βc

Critical temperature of the 2-dimensional infinite lattice Ising model determined by Onsager.

SquareIsingModel.adjacency_matrix — Function

adjacency_matrix(L, K = L)

Adjacency matrix of a L x K rectangular grid lattice.

SquareIsingModel.adjacency_tensor — Function

adjacency_tensor(L, K = L)

Adjacency tensor of a L x K rectangular grid lattice.

SquareIsingModel.binary — Method

binary(s)

Returns true or false, from different representations of a spin.

SquareIsingModel.distance_matrix — Function

distance_matrix(L, K = L)
distance_matrix(spins)

Returns a LxK matrix dist, such that the entry dist[i,j] gives the distance between sites (0,0) and (i,j) in the lattice. Takes into account the periodicity of the lattice.

SquareIsingModel.distance_tensor — Function

distance_tensor(L, K = L)

Returns a LxKxLxK tensor dist, such that the entry dist[i1,j1,i2,j2] gives the distance between sites (i1,j1) and (i2,j2) in the lattice.

SquareIsingModel.dynamic_hybrid! — Method

dynamic_hybrid!(spins, β; steps, save_interval)

Same as hybrid!, but adjusts numbers of Metropolis and Wolff steps dynamically.

SquareIsingModel.energy — Method

energy(spins, h = 0)

Computes the energy of spins in the Ising 2-dimensional model.

SquareIsingModel.hybrid! — Method

hybrid!(spins, β; steps = 1, save_interval = length(spins), local_steps = length(spins))

Hybrid sampler, performing local_steps of Metropolis sampling, then one Wolff cluster move, then another local_steps of Metropolis sampling, one more Wolff cluster move, and so on.

SquareIsingModel.ising — Method

ising(s)

Returns +1 or -1, from different representations of a spin.

SquareIsingModel.kramers_wannier — Method

kramers_wannier(β)

Returns the Kramers-Wannier dual inverse-temperature of β.

SquareIsingModel.metropolis! — Function

metropolis!(spins, β, h = 0; steps = 1, save_interval = length(spins))

Perfoms one or more Metropolis MC steps from the configuration spins, at inverse temperature β. Returns three lists: spins_t, M, E, where spins_t contains configurations sampled at intervals save_interval (by default equals the number of sites), M is the record of magnetizations, and E the record of energies.

SquareIsingModel.metropolis_acceptance_probabilities — Function

metropolis_acceptance_probabilities(β, h = 0)

Returns a tuple of acceptance probabilities, Paccept, such that if the sum of neighboring spins is S, then Paccept[(S + 4) ÷ 2 + 1] == exp(-β * ΔE), where ΔE = 2 * s[i] * (S + h) > 0 is the energy cost of flipping spin i from its current value s[i] to -s[i]. We assume that ΔE > 0, since this is the only case in which this function is called in the Metropolis algorithm. Note that S can only take the values -4, -2, 0, 2, 4, and therefore (S + 4) ÷ 2 + 1takes the values(1, 2, 3, 4, 5)`.

SquareIsingModel.metropolis_f! — Function

metropolis_f!(spins, β, h = 0; f, steps = 1, save_interval = length(spins))

Like metropolis!, but takes an additional argument f, which is a function of the magnetization that gets added to the energy, E = (...) + f(M).

SquareIsingModel.neighbor_sum — Method

neighbor_sum(spins, i, j)

Sum of spins in neighbor sites of (i, j).

SquareIsingModel.neighbor_sum_div_2 — Method

neighbor_sum_div_2(spins, i, j)

neighbor_sum(spins, i, j) is always even. Therefore here we compute the sum, divided by 2.

SquareIsingModel.neighbors — Method

neighbors(L, K [= L], i, j)
neighbors(spins, i, j)

Returns the list of neighbors of site (i,j) in the 2-dimensional lattice grid with sides L x K. That is, the sites ((i+1,j), (i-1,j), (i,j+1), (i,j-1)), but considering the periodic bounary conditions at the edges.

SquareIsingModel.onsager_heat_capacity — Method

onsager_heat_capacity(β)

Specific heat capacity of the two-dimensional Ising model, derived by Onsager in the thermodynamic limit.

SquareIsingModel.onsager_internal_energy — Method

onsager_internal_energy(β)

Internal energy per site of the two-dimensional Ising model, derived by Onsager in the thermodynamic limit.

SquareIsingModel.onsager_magnetization — Method

onsager_magnetization(β)

Analytical magnetization of the two-dimensional Ising model found by Onsager in the thermodynamic limit.

SquareIsingModel.random_configuration — Method

random_configuration(L, K = L)

Generate a random spin configuration.

SquareIsingModel.random_magnetized_configuration — Function

random_magnetized_configuration(M, L, K = L)

Generate a random spin configuration with magnetization (sum over all spins) equal to M. You must make sure that the given value M is feasible.

SquareIsingModel.wolff! — Method

wolff!(spins, β; steps = 1, save_interval = 1)

Perfoms one or more Wolff MC steps from the configuration spins, at inverse temperature β.