Reference

This package doesn't export any symbols.

Grid lattice

SquareIsingModel.energy — Function

energy(spins)

Computes the energy of spins in the Ising 2-dimensional model.

SquareIsingModel.adjacency_matrix — Function

adjacency_matrix(L, K = L)

Adjacency matrix of a L x K rectangular grid lattice.

SquareIsingModel.neighbors — Function

neighbors(i, j, L, K = L)

Returns the list of neighbors of site (i,j) in the 2-dimensional lattice grid with sides L x K. That is, the sites ((i+1,j), (i-1,j), (i,j+1), (i,j-1)), but considering the periodic bounary conditions at the edges.

SquareIsingModel.neighbor_sum — Function

neighbor_sum(spins, i, j)

Sum of spins in neighbor sites of (i, j).

SquareIsingModel.random_configuration — Function

random_configuration(L, K = L)

Generate a random spin configuration.

Onsager analytical solution

SquareIsingModel.βc — Constant

βc

Critical temperature of the 2-dimensional infinite lattice Ising model determined by Onsager.

SquareIsingModel.onsager_magnetization — Function

onsager_magnetization(β)

Analytical magnetization of the two-dimensional Ising model found by Onsager in the thermodynamic limit.

Monte Carlo simulations

SquareIsingModel.metropolis! — Function

metropolis!(spins, β, steps = 1; save_interval = length(spins))

Perfoms one or more Metropolis MC steps from the configuration spins, at inverse temperature β. Returns three lists: spins_t, M, E, where spins_t contains configurations sampled at intervals save_interval (by default equals the number of sites), M is the record of magnetizations, and E the record of energies.

SquareIsingModel.wolff! — Function

wolff!(spins, β, steps = 1; save_interval = 1)

Perfoms one or more Wolff MC steps from the configuration spins, at inverse temperature β.